Introduce and use FingerprintDeserializer
