+++ tests/unit/DeserializerFactoryTest.php
@@ -120,4 +120,11 @@ class DeserializerFactoryTest extends \PHPUnit_Framework_TestCase {
 		);
 	}
 
+	public function testNewFingerprintDeserializer() {
+		$this->assertDeserializesWithoutException(
+			$this->buildDeserializerFactory()->newFingerprintDeserializer(),
+			array()
+		);
+	}
+
 }
